refactor: version history repository
