+++ server/src/repositories/version-history.repository.ts
@@ -3,25 +3,23 @@ import { Insertable, Kysely } from 'kysely';
 import { InjectKysely } from 'nestjs-kysely';
 import { DB, VersionHistory } from 'src/db';
 import { GenerateSql } from 'src/decorators';
-import { VersionHistoryEntity } from 'src/entities/version-history.entity';
-import { IVersionHistoryRepository } from 'src/interfaces/version-history.interface';
 
 @Injectable()
-export class VersionHistoryRepository implements IVersionHistoryRepository {
+export class VersionHistoryRepository {
   constructor(@InjectKysely() private db: Kysely<DB>) {}
 
   @GenerateSql()
-  getAll(): Promise<VersionHistoryEntity[]> {
+  getAll() {
     return this.db.selectFrom('version_history').selectAll().orderBy('createdAt', 'desc').execute();
   }
 
   @GenerateSql()
-  getLatest(): Promise<VersionHistoryEntity | undefined> {
+  getLatest() {
     return this.db.selectFrom('version_history').selectAll().orderBy('createdAt', 'desc').executeTakeFirst();
   }
 
   @GenerateSql({ params: [{ version: 'v1.123.0' }] })
-  create(version: Insertable<VersionHistory>): Promise<VersionHistoryEntity> {
+  create(version: Insertable<VersionHistory>) {
     return this.db.insertInto('version_history').values(version).returningAll().executeTakeFirstOrThrow();
   }
 }
